Rename base package from `org.example.demo` to `org.example.untitled`.
