+++ src/main/java/org/example/untitled/DemoApplication.java
@@ -1,4 +1,4 @@
-package org.example.demo;
+package org.example.untitled;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
