+++ src/main/java/org/example/untitled/entity/CaseStatus.java
@@ -1,4 +1,4 @@
-package org.example.demo.entity;
+package org.example.untitled.entity;
 
 public enum CaseStatus {
     OPEN,
